src/requirements.txt
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0